+++ RELEASE_NOTES.md
@@ -1,1 +1,7 @@
-Release note for version 5.1.1 not found.
+Bugfix release
+
+**Bugfixes**
+- Fix update_4_to_5.0.0.sql script
+
+**Improvements**
+- Improve database repair process to find key case mismatches